Adding data and tables file
This Excel sheet comprises the data and tables associated with "Exploring the association of weather variability on Campylobacter - a systematic review" published 2024 by Austhof et al. in Environmental Research. These tables and associated data are available for others to use as long as the above DOI and paper are cited appropriately.